:tada: 3.2.0.RELEASE 新增灵活数据权限特性

+++ blade-service/blade-desk/src/test/java/BladeDemoTest.java
@@ -22,7 +22,7 @@ public class BladeDemoTest {
 
 	@Test
 	public void contextLoads() {
-		int count = noticeService.count();
+		Long count = noticeService.count();
 		System.out.println("notice数量：[" + count + "] 个");
 	}
